Version 0.2.0 update
This includes a pretty big overhall where I've restyled all of the functions to be snake_case rather than BigCamelCase. This was for a few reasons, but primarily because it matches the style of the tidyverse, which this project uses. Incomplete functions, were made complete and the previous versions have changed their formatting without depreciation. Completed functions have been depreciated.

I've also removed the packrat add in. I found that it was creating more work than it was worth as this is a very lightweight package to aid those who work with the FWF from the ACT.